--- app/views/articles/_tabs.html.haml
@@ -1,5 +1,0 @@
-.tabs
-  .tabs_item.active
-    新着記事一覧
-  .tabs_item
-    タイムライン
