+++ app/views/commons/_articles.html.haml
@@ -1,0 +1,7 @@
+.container
+  = render 'commons/tabs'
+  - articles.each do |article|
+    = render 'commons/article', article: article
+  = link_to new_article_path do
+    .float_btn
+      %i.fa.fa-plus
