+++ config/routes.rb
@@ -3,6 +3,8 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   root to: 'articles#index'
 
+  resource :timeline, only: [:show]
+
   resources :articles do
     resources :comments, only: [:new, :create]
